requirements.txt
# Python dependencies for the Sales Quotation Bot

# Core LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Database dependencies (using your existing setup)
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Environment and configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Async support
asyncio-mqtt>=0.11.0

# Optional: For enhanced AI capabilities
# openai>=1.0.0
# anthropic>=0.3.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
